+++ web/src/compose_call_ui.ts
@@ -174,6 +174,29 @@ export function generate_and_insert_audio_or_video_call_link(
 
                 break;
             }
+            case available_providers.nextcloud_talk?.id: {
+                const room_name = `${get_recipient_label()?.label_text ?? ""} conversation`;
+                const request = {room_name};
+
+                channel.post({
+                    url: "/json/calls/nextcloud_talk/create",
+                    data: request,
+                    success(response) {
+                        const data = call_response_schema.parse(response);
+                        insert_video_call_url(data.url, $target_textarea);
+                    },
+                    error(_, status) {
+                        if (status !== "abort") {
+                            ui_report.generic_embed_error(
+                                $t_html({defaultMessage: "Failed to create video call."}),
+                                2000,
+                            );
+                        }
+                    },
+                });
+
+                break;
+            }
             case available_providers.jitsi_meet?.id: {
                 const video_call_id = util.random_int(100000000000000, 999999999999999);
                 const video_call_url = compose_call.get_jitsi_server_url(video_call_id.toString());