+++ web/tests/settings_org.test.cjs
@@ -443,6 +443,10 @@ test("set_up", ({override, override_rewire}) => {
             id: 6,
             name: "Constructor Groups",
         },
+        nextcloud_talk: {
+            id: 7,
+            name: "Nextcloud Talk",
+        },
     });
     override(realm, "realm_message_retention_days", null);