+++ zproject/urls.py
@@ -269,6 +269,7 @@ from zerver.views.users import (
 )
 from zerver.views.video_calls import (
     complete_zoom_user,
+    create_nextcloud_talk_url,
     deauthorize_zoom_user,
     get_bigbluebutton_url,
     join_bigbluebutton,
@@ -603,6 +604,8 @@ v1_api_and_json_patterns = [
     rest_path("calls/bigbluebutton/create", GET=get_bigbluebutton_url),
     # Used to generate a Constructor Groups video call URL
     rest_path("calls/constructorgroups/create", POST=make_constructor_groups_video_call),
+    # Used to generate a Nextcloud Talk video call URL
+    rest_path("calls/nextcloud_talk/create", POST=create_nextcloud_talk_url),
     # export/realm -> zerver.views.realm_export
     rest_path("export/realm", POST=export_realm, GET=get_realm_exports),
     rest_path("export/realm/<int:export_id>", DELETE=delete_realm_export),